14.0 작업ㅣ프레임워크(Exiled)가 정상화되기를 기다리는 중 (1/20)

+++ RGM/RGM/Modes/Pluses/Outlaw/Outlaw.cs
@@ -79,7 +79,7 @@ namespace RGM.Modes
                     if (firearm.AmmoType != AmmoType.None)
                     {
                         for (int i = 0; i < 3; i++)
-                            player.AddAmmo(firearm.AmmoType, firearm.MaxAmmo);
+                            player.AddAmmo(firearm.AmmoType, (ushort)firearm.MaxAmmo);
                     }
                 }
             }
